+++ AmperfyTests/Cases/SongTest.swift
@@ -37,7 +37,7 @@ class SongTest: XCTestCase {
     }
     
     func testArtist() {
-        guard let artist = storage.getArtist(id: Int32(cdHelper.seeder.artists[0].id)) else { XCTFail(); return }
+        guard let artist = storage.getArtist(id: cdHelper.seeder.artists[0].id) else { XCTFail(); return }
         testSong.artist = artist
         XCTAssertEqual(testSong.artist!.id, artist.id)
         storage.saveContext()
